requirements.txt
python-telegram-bot==13.15
pymorphy2
emoji
